# Cache module resolutions across tsbuild to be able to resolve the modules faster