This is first of many optimizations done for building 800+ reference projects
cc @MLoughry as its coming from the repro he provided

Build | release-3.4 | this branch of release-3.4
------------ | -------------|------------
Graph calculation | 16.421s | 14.296s
Building Project 0/802 | 3.266s | 2.985s
Elapsed | 19.687s | 17.281s
Building Project 1/802 | 0.672s | 0.297s
Elapsed | 20.359s | 17.578s
Building Project 2/802 | 0.578s | 0.25s
Elapsed | 20.937s | 17.828s
Building Project 3/802 | 1.188s | 0.828s
Elapsed | 22.125s | 18.656s
Building Project 4/802 | 0.515s | 0.234s
Elapsed | 22.64s | 18.89s
Building Project 5/802 | 0.578s | 0.281s
Elapsed | 23.218s | 19.171s
Building Project 6/802 | 0.532s | 0.219s
Elapsed | 23.75s | 19.39s
Building Project 7/802 | 1.015s | 0.656s
Elapsed | 24.765s | 20.046s
Building Project 8/802 | 1.203s | 1.125s
Elapsed | 25.968s | 21.171s
Building Project 9/802 | 1.141s | 0.625s
Elapsed | 27.109s | 21.796s
Building Project 10/802 | 1s | 0.485s
Elapsed | 28.109s | 22.281s
Building Project 11/802 | 0.609s | 0.312s
Elapsed | 28.718s | 22.593s
Building Project 12/802 | 0.75s | 0.297s
Elapsed | 29.468s | 22.89s
Building Project 13/802 | 0.563s | 0.281s
Elapsed | 30.031s | 23.171s
Building Project 14/802 | 0.562s | 0.266s
Elapsed | 30.593s | 23.437s
Building Project 15/802 | 0.563s | 0.281s
Elapsed | 31.156s | 23.718s
Building Project 16/802 | 0.562s | 0.266s
Elapsed | 31.718s | 23.984s
Building Project 17/802 | 0.578s | 0.281s
Elapsed | 32.296s | 24.265s
Building Project 18/802 | 0.859s | 0.531s
Elapsed | 33.171s | 24.796s
Building Project 19/802 | 0.953s | 0.563s
Elapsed | 34.124s | 25.359s
Building Project 20/802 | 0.657s | 0.265s
Elapsed | 34.781s | 25.624s
Building Project 21/802 | 0.484s | 0.203s
Elapsed | 35.265s | 25.843s
Building Project 22/802 | 0.531s | 0.219s
Elapsed | 35.796s | 26.062s
Building Project 23/802 | 0.5s | 0.203s
Elapsed | 36.296s | 26.265s
Building Project 24/802 | 0.5s | 0.203s
Elapsed | 36.796s | 26.468s
Building Project 25/802 | 1.282s | 0.828s
Elapsed | 38.093s | 27.296s
Building Project 26/802 | 0.703s | 0.36s
Elapsed | 38.796s | 27.656s
Building Project 27/802 | 0.5s | 0.203s
Elapsed | 39.296s | 27.859s
Building Project 28/802 | 0.532s | 0.219s
Elapsed | 39.828s | 28.078s
Building Project 29/802 | 0.797s | 0.203s
Elapsed | 40.625s | 28.281s
Building Project 30/802 | 1.031s | 0.609s
Elapsed | 41.656s | 28.89s
Building Project 31/802 | 0.719s | 0.359s
Elapsed | 42.375s | 29.249s
Building Project 32/802 | 10.593s | 2.75s
Elapsed | 52.968s | 31.999s
Building Project 33/802 | 0.829s | 0.5s
Elapsed | 53.797s | 32.499s
Building Project 34/802 | 0.546s | 0.25s
Elapsed | 54.343s | 32.749s
Building Project 35/802 | 1.032s | 0.204s
Elapsed | 55.375s | 32.953s
Building Project 36/802 | 0.813s | 0.234s
Elapsed | 56.203s | 33.187s
Building Project 37/802 | 0.656s | 0.359s
Elapsed | 56.859s | 33.546s
Building Project 38/802 | 1.109s | 0.735s
Elapsed | 57.968s | 34.281s
Building Project 39/802 | 1.063s | 0.64s
Elapsed | 59.031s | 34.921s
Building Project 40/802 | 0.625s | 0.344s
Elapsed | 59.656s | 35.265s
Building Project 41/802 | 3.843s | 1.156s
Elapsed | 63.499s | 36.421s
Building Project 42/802 | 0.984s | 0.297s
Elapsed | 64.483s | 36.718s
Building Project 43/802 | 3.125s | 1.641s
Elapsed | 67.608s | 38.359s
Building Project 44/802 | 0.766s | 0.422s
Elapsed | 68.374s | 38.781s
Building Project 45/802 | 0.89s | 0.594s
Elapsed | 69.28s | 39.39s
Building Project 46/802 | 2.297s | 1.813s
Elapsed | 71.577s | 41.203s
Building Project 47/802 | 1.172s | 0.765s
Elapsed | 72.749s | 41.968s
Building Project 48/802 | 1.89s | 1.453s
Elapsed | 74.639s | 43.421s
Building Project 49/802 | 0.766s | 0.36s
Elapsed | 75.405s | 43.781s
Building Project 50/802 | 0.5s | 0.203s
Elapsed | 75.905s | 43.984s
Building Project 51/802 | 0.516s | 0.219s
Elapsed | 76.421s | 44.203s
Building Project 52/802 | 0.562s | 0.25s
Elapsed | 76.983s | 44.453s
Building Project 53/802 | 1.047s | 0.625s
Elapsed | 78.03s | 45.093s
Building Project 54/802 | 1.75s | 0.766s
Elapsed | 79.78s | 45.859s
Building Project 55/802 | 0.516s | 0.204s
Elapsed | 80.296s | 46.078s
Building Project 56/802 | 0.625s | 0.281s
Elapsed | 80.921s | 46.359s
Building Project 57/802 | 2.296s | 1.437s
Elapsed | 83.217s | 47.796s
Building Project 58/802 | 0.485s | 0.219s
Elapsed | 83.702s | 48.015s
Building Project 59/802 | 1.187s | 0.625s
Elapsed | 84.889s | 48.64s
Building Project 60/802 | 2.578s | 1.984s
Elapsed | 87.467s | 50.624s
Building Project 61/802 | 2.172s | 1.829s
Elapsed | 89.639s | 52.453s
Building Project 62/802 | 1.532s | 1.14s
Elapsed | 91.171s | 53.593s
Building Project 63/802 | 1.546s | 1.141s
Elapsed | 92.717s | 54.734s
Building Project 64/802 | 1.766s | 1.234s
Elapsed | 94.483s | 55.968s
Building Project 65/802 | 0.609s | 0.344s
Elapsed | 95.092s | 56.312s
Building Project 66/802 | 0.813s | 0.437s
Elapsed | 95.905s | 56.749s
Building Project 67/802 | 0.453s | 0.188s
Elapsed | 96.358s | 56.937s
Building Project 68/802 | 0.703s | 0.359s
Elapsed | 97.061s | 57.296s
Building Project 69/802 | 0.578s | 0.297s
Elapsed | 97.639s | 57.593s
Building Project 70/802 | 0.75s | 0.422s
Elapsed | 98.389s | 58.015s
Building Project 71/802 | 0.828s | 0.406s
Elapsed | 99.217s | 58.421s
Building Project 72/802 | 1.063s | 0.625s
Elapsed | 100.28s | 59.046s
Building Project 73/802 | 0.687s | 0.375s
Elapsed | 100.967s | 59.421s
Building Project 74/802 | 1.735s | 1.265s
Elapsed | 102.702s | 60.686s
Building Project 75/802 | 2.969s | 2.406s
Elapsed | 105.671s | 63.092s
Building Project 76/802 | 1.531s | 1.078s
Elapsed | 107.202s | 64.17s
Building Project 77/802 | 0.937s | 0.578s
Elapsed | 108.139s | 64.748s
Building Project 78/802 | 1.906s | 1.094s
Elapsed | 110.061s | 65.842s
Building Project 79/802 | 0.61s | 0.656s
Elapsed | 110.671s | 66.498s
Building Project 80/802 | 0.515s | 0.235s
Elapsed | 111.186s | 66.748s
Building Project 81/802 | 0.516s | 0.235s
Elapsed | 111.702s | 66.983s
Building Project 82/802 | 0.859s | 0.453s
Elapsed | 112.561s | 67.436s
Building Project 83/802 | 0.485s | 0.188s
Elapsed | 113.046s | 67.624s
Building Project 84/802 | 0.906s | 0.562s
Elapsed | 113.952s | 68.186s
Building Project 85/802 | 1.312s | 0.906s
Elapsed | 115.264s | 69.092s
Building Project 86/802 | 2.188s | 1.656s
Elapsed | 117.452s | 70.748s
Building Project 87/802 | 0.484s | 0.219s
Elapsed | 117.936s | 70.967s
Building Project 88/802 | 0.719s | 0.391s
Elapsed | 118.655s | 71.358s
Building Project 89/802 | 2.078s | 1.25s
Elapsed | 120.733s | 72.608s
Building Project 90/802 | 0.922s | 0.594s
Elapsed | 121.655s | 73.202s
Building Project 91/802 | 0.5s | 0.234s
Elapsed | 122.155s | 73.436s
Building Project 92/802 | 1.67s | 1.187s
Elapsed | 123.825s | 74.623s
Building Project 93/802 | 1.859s | 1.875s
Elapsed | 125.7s | 76.498s
Building Project 94/802 | 0.765s | 0.438s
Elapsed | 126.465s | 76.936s
Building Project 95/802 | 0.485s | 0.234s
Elapsed | 126.95s | 77.17s
Building Project 96/802 | 1.14s | 0.828s
Elapsed | 128.106s | 77.998s
Building Project 97/802 | 0.547s | 0.266s
Elapsed | 128.653s | 78.264s
Building Project 98/802 | 2.547s | 1.656s
Elapsed | 131.2s | 79.92s
Building Project 99/802 | 2.172s | 1.579s
Elapsed | 133.372s | 81.499s
Building Project 100/802 | 0.718s | 0.359s
Elapsed | 134.09s | 81.858s
Building Project 101/802 | 0.719s | 0.39s
Elapsed | 134.809s | 82.248s
Building Project 102/802 | 2.61s | 2.047s
Elapsed | 137.419s | 84.295s
Building Project 103/802 | 3.546s | 2.469s
Elapsed | 140.965s | 86.78s
Building Project 104/802 | 2.563s | 2.312s
Elapsed | 143.528s | 89.108s
Building Project 105/802 | 1.719s | 1.344s
Elapsed | 145.247s | 90.452s
Building Project 106/802 | 0.75s | 0.421s
Elapsed | 145.997s | 90.873s
Building Project 107/802 | 0.843s | 0.454s
Elapsed | 146.84s | 91.342s
Building Project 108/802 | 0.719s | 0.36s
Elapsed | 147.559s | 91.702s
Building Project 109/802 | 0.719s | 0.328s
Elapsed | 148.278s | 92.03s
Building Project 110/802 | 3.437s | 2.5s
Elapsed | 151.715s | 94.53s
Building Project 111/802 | 2.86s | 1.922s
Elapsed | 154.575s | 96.452s
Building Project 112/802 | 3.922s | 2.781s
Elapsed | 158.497s | 99.233s
Building Project 113/802 | 2.375s | 1.765s
Elapsed | 160.872s | 100.998s
Building Project 114/802 | 1s | 0.625s
Elapsed | 161.872s | 101.623s
Building Project 115/802 | 2.547s | 1.922s
Elapsed | 164.419s | 103.545s
Building Project 116/802 | 0.75s | 0.422s
Elapsed | 165.184s | 103.967s
Building Project 117/802 | 2.719s | 1.75s
Elapsed | 167.903s | 105.717s
Building Project 118/802 | 0.687s | 0.328s
Elapsed | 168.59s | 106.045s
Building Project 119/802 | 0.469s | 0.204s
Elapsed | 169.075s | 106.264s
Building Project 120/802 | 0.484s | 0.219s
Elapsed | 169.559s | 106.483s
Building Project 121/802 | 0.625s | 0.312s
Elapsed | 170.184s | 106.795s
Building Project 122/802 | 1.516s | 1.016s
Elapsed | 171.7s | 107.811s
Building Project 123/802 | 1.609s | 1.594s
Elapsed | 173.309s | 109.405s
Building Project 124/802 | 2.219s | 1.734s
Elapsed | 175.528s | 111.139s
Building Project 125/802 | 2.094s | 1.61s
Elapsed | 177.622s | 112.749s
Building Project 126/802 | 2.218s | 1.687s
Elapsed | 179.84s | 114.436s
Building Project 127/802 | 1.922s | 0.938s
Elapsed | 181.762s | 115.374s
Building Project 128/802 | 0.563s | 0.25s
Elapsed | 182.325s | 115.624s
Building Project 129/802 | 0.922s | 0.578s
Elapsed | 183.247s | 116.202s
Building Project 130/802 | 1.201s | 0.796s
Elapsed | 184.448s | 116.998s
Building Project 131/802 | 2.89s | 2.735s
Elapsed | 187.354s | 119.733s
Building Project 132/802 | 0.75s | 0.515s
Elapsed | 188.104s | 120.248s
Building Project 133/802 | 1.625s | 1.249s
Elapsed | 189.729s | 121.497s
Building Project 134/802 | 2.953s | 2.344s
Elapsed | 192.682s | 123.841s
Building Project 135/802 | 2.969s | 2.25s
Elapsed | 195.651s | 126.091s
Building Project 136/802 | 1.469s | 1.078s
Elapsed | 197.12s | 127.169s
Building Project 137/802 | 5.843s | 4.703s
Elapsed | 202.963s | 131.872s
Building Project 138/802 | 1.141s | 0.704s
Elapsed | 204.104s | 132.576s
Building Project 139/802 | 1.047s | 0.625s
Elapsed | 205.151s | 133.201s
Building Project 140/802 | 0.593s | 0.25s
Elapsed | 205.76s | 133.451s
Building Project 141/802 | 1.641s | 0.765s
Elapsed | 207.401s | 134.216s
Building Project 142/802 | 4.578s | 3.094s
Elapsed | 211.979s | 137.326s
Building Project 143/802 | 2.875s | 2s
Elapsed | 214.854s | 139.326s
Building Project 144/802 | 1.781s | 1.125s
Elapsed | 216.635s | 140.466s
Building Project 145/802 | 1.36s | 0.875s
Elapsed | 217.995s | 141.357s
Building Project 146/802 | 2.625s | 1.5s
Elapsed | 220.62s | 142.873s
Building Project 147/802 | 1.703s | 1.249s
Elapsed | 222.338s | 144.122s
Building Project 148/802 | 0.563s | 0.235s
Elapsed | 222.901s | 144.357s
Building Project 149/802 | 1.047s | 0.719s
Elapsed | 223.948s | 145.076s
Building Project 150/802 | 0.875s | 0.609s
Elapsed | 224.823s | 145.685s
Building Project 151/802 | 2.594s | 2.156s
Elapsed | 227.432s | 147.841s
Building Project 152/802 | 4.156s | 2.922s
Elapsed | 231.588s | 150.763s
Building Project 153/802 | 2.547s | 2.203s
Elapsed | 234.135s | 152.966s
Building Project 154/802 | 3.235s | 1.953s
Elapsed | 237.37s | 154.919s
Building Project 155/802 | 3.797s | 2.672s
Elapsed | 241.167s | 157.591s
Building Project 156/802 | 3.451s | 3.031s
Elapsed | 244.633s | 160.622s
Building Project 157/802 | 4.032s | 3.094s
Elapsed | 248.665s | 163.716s
Building Project 158/802 | 3.093s | 2.563s
Elapsed | 251.758s | 166.279s
Building Project 159/802 | 1.641s | 1.156s
Elapsed | 253.399s | 167.435s
Building Project 160/802 | 3.953s | 2.672s
Elapsed | 257.352s | 170.122s
Building Project 161/802 | 1.047s | 0.672s
Elapsed | 258.399s | 170.794s
Building Project 162/802 | 2.219s | 1.625s
Elapsed | 260.633s | 172.419s
Building Project 163/802 | 2.094s | 1.563s
Elapsed | 262.727s | 173.982s
Building Project 164/802 | 2.141s | 1.578s
Elapsed | 264.883s | 175.576s
Building Project 165/802 | 0.641s | 0.343s
Elapsed | 265.524s | 175.919s
Building Project 166/802 | 0.968s | 0.61s
Elapsed | 266.508s | 176.529s
Building Project 167/802 | 3.782s | 2.437s
Elapsed | 270.29s | 178.966s
Building Project 168/802 | 2.296s | 1.75s
Elapsed | 272.586s | 180.716s
Building Project 169/802 | 2.157s | 2.139s
Elapsed | 274.743s | 182.871s
Building Project 170/802 | 3s | 2.297s
Elapsed | 277.743s | 185.184s
Building Project 171/802 | 3.562s | 2.156s
Elapsed | 281.305s | 187.34s
Building Project 172/802 | 2.469s | 1.766s
Elapsed | 283.774s | 189.106s
Building Project 173/802 | 2.109s | 1.562s
Elapsed | 285.883s | 190.668s
Building Project 174/802 | 2.672s | 2.516s
Elapsed | 288.555s | 193.184s
Building Project 175/802 | 1.188s | 0.875s
Elapsed | 289.743s | 194.059s
Building Project 176/802 | 1.015s | 0.687s
Elapsed | 290.758s | 194.746s
Building Project 177/802 | 2.907s | 2s
Elapsed | 293.665s | 196.746s
Building Project 178/802 | 2.484s | 1.876s
Elapsed | 296.149s | 198.622s
Building Project 179/802 | 2.719s | 2.031s
Elapsed | 298.883s | 200.668s
Building Project 180/802 | 1.375s | 0.844s
Elapsed | 300.258s | 201.512s
Building Project 181/802 | 3.235s | 2.547s
Elapsed | 303.508s | 204.059s
Building Project 182/802 | 1.999s | 1.406s
Elapsed | 305.507s | 205.465s
Building Project 183/802 | 2.375s | 1.766s
Elapsed | 307.882s | 207.231s
Building Project 184/802 | 2.282s | 1.765s
Elapsed | 310.164s | 208.996s
Building Project 185/802 | 2.14s | 1.641s
Elapsed | 312.304s | 210.637s
Building Project 186/802 | 2.281s | 2.11s
Elapsed | 314.617s | 212.747s
Building Project 187/802 | 2.812s | 1.703s
Elapsed | 317.429s | 214.465s
Building Project 188/802 | 3.656s | 2.532s
Elapsed | 321.085s | 216.997s
Building Project 189/802 | 1.829s | 1.359s
Elapsed | 322.914s | 218.356s
Building Project 190/802 | 3.39s | 3.047s
Elapsed | 326.304s | 221.403s
Building Project 191/802 | 1.516s | 1.547s
Elapsed | 327.82s | 222.95s
Building Project 192/802 | 0.75s | 0.39s
Elapsed | 328.57s | 223.34s
Building Project 193/802 | 0.64s | 0.313s
Elapsed | 329.21s | 223.653s
Building Project 194/802 | 2.204s | 1.609s
Elapsed | 331.414s | 225.262s
Building Project 195/802 | 1.875s | 0.938s
Elapsed | 333.289s | 226.2s
Building Project 196/802 | 1.406s | 0.937s
Elapsed | 334.695s | 227.137s
Building Project 197/802 | 4.672s | 3.578s
Elapsed | 339.367s | 230.715s
Building Project 198/802 | 1.453s | 1.516s
Elapsed | 340.82s | 232.231s
Building Project 199/802 | 2.156s | 1.703s
Elapsed | 342.976s | 233.934s
Building Project 200/802 | 2.422s | 1.859s
Elapsed | 345.398s | 235.793s
Building Project 201/802 | 1.937s | 1.454s
Elapsed | 347.335s | 237.247s
Building Project 202/802 | 2.204s | 1.625s
Elapsed | 349.539s | 238.872s
Building Project 203/802 | 2.718s | 2.093s
Elapsed | 352.257s | 240.965s
Building Project 204/802 | 2.657s | 2.029s
Elapsed | 354.914s | 242.994s
Building Project 205/802 | 3.625s | 2.875s
Elapsed | 358.539s | 245.869s
Building Project 206/802 | 1.578s | 1.14s
Elapsed | 360.117s | 247.009s
Building Project 207/802 | 2.281s | 1.703s
Elapsed | 362.398s | 248.712s
Building Project 208/802 | 2.078s | 1.594s
Elapsed | 364.476s | 250.306s
Building Project 209/802 | 2.904s | 1.859s
Elapsed | 367.38s | 252.165s
Building Project 210/802 | 2.188s | 1.703s
Elapsed | 369.583s | 253.9s
Building Project 211/802 | 2.172s | 1.656s
Elapsed | 371.755s | 255.556s
Building Project 212/802 | 0.485s | 0.234s
Elapsed | 372.24s | 255.79s
Building Project 213/802 | 0.937s | 0.579s
Elapsed | 373.177s | 256.384s
Building Project 214/802 | 1.016s | 0.703s
Elapsed | 374.193s | 257.087s
Building Project 215/802 | 0.828s | 0.547s
Elapsed | 375.021s | 257.634s
Building Project 216/802 | 2.156s | 2.235s
Elapsed | 377.177s | 259.869s
Building Project 217/802 | 2.641s | 2.156s
Elapsed | 379.818s | 262.025s
Building Project 218/802 | 7.562s | 6.422s
Elapsed | 387.396s | 268.462s
Building Project 219/802 | 2.937s | 2.297s
Elapsed | 390.333s | 270.759s
Building Project 220/802 | 1.485s | 1.031s
Elapsed | 391.818s | 271.79s
Building Project 221/802 | 1.875s | 1.375s
Elapsed | 393.693s | 273.165s
Building Project 222/802 | 3.734s | 2.688s
Elapsed | 397.427s | 275.853s
Building Project 223/802 | 1.469s | 0.578s
Elapsed | 398.896s | 276.447s
Building Project 224/802 | 6s | 4.843s
Elapsed | 404.896s | 281.29s
Building Project 225/802 | 2.047s | 1.625s
Elapsed | 406.943s | 282.915s
Building Project 226/802 | 1.265s | 0.875s
Elapsed | 408.208s | 283.79s
Building Project 227/802 | 3.079s | 2.704s
Elapsed | 411.287s | 286.494s
Building Project 228/802 | 0.609s | 0.281s
Elapsed | 411.896s | 286.79s
Building Project 229/802 | 2.187s | 1.672s
Elapsed | 414.083s | 288.462s
Building Project 230/802 | 2.25s | 1.703s
Elapsed | 416.333s | 290.165s
Building Project 231/802 | 2.485s | 1.922s
Elapsed | 418.818s | 292.087s
Building Project 232/802 | 2.156s | 1.594s
Elapsed | 420.974s | 293.681s
Building Project 233/802 | 2.078s | 1.578s
Elapsed | 423.052s | 295.259s
Building Project 234/802 | 0.735s | 0.36s
Elapsed | 423.787s | 295.619s
Building Project 235/802 | 2.234s | 1.703s
Elapsed | 426.021s | 297.322s
Building Project 236/802 | 0.641s | 0.343s
Elapsed | 426.662s | 297.665s
Building Project 237/802 | 1.529s | 1.141s
Elapsed | 428.191s | 298.806s
Building Project 238/802 | 0.984s | 0.625s
Elapsed | 429.175s | 299.431s
Building Project 239/802 | 0.563s | 0.281s
Elapsed | 429.738s | 299.712s
Building Project 240/802 | 1.156s | 0.782s
Elapsed | 430.894s | 300.494s
Building Project 241/802 | 2.453s | 1.781s
Elapsed | 433.347s | 302.275s
Building Project 242/802 | 0.734s | 0.453s
Elapsed | 434.081s | 302.728s
Building Project 243/802 | 5.704s | 4.733s
Elapsed | 439.785s | 307.461s
Building Project 244/802 | 3.172s | 2.078s
Elapsed | 442.972s | 309.539s
Building Project 245/802 | 2.406s | 1.938s
Elapsed | 445.378s | 311.477s
Building Project 246/802 | 1.985s | 2.078s
Elapsed | 447.363s | 313.555s
Building Project 247/802 | 2.718s | 2.281s
Elapsed | 450.081s | 315.836s
Building Project 248/802 | 2.032s | 1.61s
Elapsed | 452.113s | 317.446s
Building Project 249/802 | 2.515s | 1.531s
Elapsed | 454.628s | 318.977s
Building Project 250/802 | 2.313s | 1.718s
Elapsed | 456.941s | 320.711s
Building Project 251/802 | 2.015s | 1.547s
Elapsed | 458.956s | 322.258s
Building Project 252/802 | 2.125s | 1.625s
Elapsed | 461.081s | 323.899s
Building Project 253/802 | 0.531s | 0.234s
Elapsed | 461.628s | 324.133s
Building Project 254/802 | 2.797s | 2.688s
Elapsed | 464.425s | 326.821s
Building Project 255/802 | 2.281s | 1.687s
Elapsed | 468.675s | 328.508s
Building Project 256/802 | 1.906s | 1.453s
Elapsed | 470.581s | 329.961s
Building Project 257/802 | 1.907s | 1.438s
Elapsed | 472.488s | 331.399s
Building Project 258/802 | 2.187s | 1.641s
Elapsed | 474.675s | 333.055s
Building Project 259/802 | 2.656s | 1.641s
Elapsed | 477.331s | 334.696s
Building Project 260/802 | 2.125s | 1.578s
Elapsed | 479.456s | 336.274s
Building Project 261/802 | 0.797s | 0.437s
Elapsed | 480.253s | 336.711s
Building Project 262/802 | 2.61s | 2.219s
Elapsed | 482.863s | 338.93s
Building Project 263/802 | 2.578s | 2.031s
Elapsed | 485.441s | 340.961s
Building Project 264/802 | 2.234s | 1.688s
Elapsed | 487.675s | 342.649s
Building Project 265/802 | 2.718s | 2s
Elapsed | 490.393s | 344.649s
Building Project 266/802 | 2.937s | 2.203s
Elapsed | 493.33s | 346.852s
Building Project 267/802 | 2.25s | 1.672s
Elapsed | 495.58s | 348.524s
Building Project 268/802 | 4.907s | 3.781s
…