@DanielRosenwasser 2019-04-26
Does this benefit `--incremental` users at all?