@sheetalkamat 2019-04-26
@DanielRosenwasser little bit but not much (if their project contained references and references had same options as the project being compiled, it will)